libcapsule is a library used to implement proxying/shim libraries
The generated shim libraries (libz and libGL are included as examples)
provide the same symbols as the target library but do not expose the
symbols from their dependencies to other objects that link to them.

They are also capable of using a filesystem tree starting somwhere
other than / to get the dependencies for the target.

Included are tools to help generate said proxy libraries with minimal
manual intervention.