Modify generated file dependencies to allow optional dependencies
We now reserve the shim/lib<FOO>.so.c.<WHATEVER> pattern for the
dependencies of the generated .so.c file.

The shim/lib<FOO>.so.<WHATEVER> pattern is now used for generated
intermediates.

This allows two useful things: We don't occasionally get a second
rebuild of the generated .so.c file due to timer resolution problems
and we can depend on extra .so.c.* files if they exist and ignore
them otherwise: we will need the latter as some libraries (libGL)
need bespoke dlopen/dlmopen wrapper logic and others can just live
with the default implementation.