ELF32_R_x vs EL64_R_x: compiler not smart enough to infer path safety
If we use the ELF64 and ELF32 _R_x macros in the relevant branches of
a switch( __ELF_NATIVE_CLASS ) case statement we're guaranteed to be
safe but the compiler isn't smart enough to deduce this and emits some
warnings about bit shift overflows.

Use #if __ELF_NATIVE_CLASS == x code selection instead.