gen-stublib.sh: Include the major version of the proxied library
When searching the ld cache, calling dlopen etc we cannot assume
that there will be a libfoo.so linkk - or that it will point to
the right major version of the library. Always search for an explcit
libfoo.so.X target.

Value of X is automatically parsed from the arguments to libtool.

X defaults to 0.