capsule-dlmopen.c: add some debug output to ELF DSO type checking
Dump out the class (ELFCLASS32 vs ELFCLASS64) and the machine
(eg EM_386 vs EM_X86_64) of candidate DSOs.

This matters when the target file tree is multi-arch, anc can contaim
different classes and machines.

x86 hardware typically has 3:

  ELFCLASS32/EM_386    - old school i[3456]86 binary
  ELFCLASS64/EM_X86_64 - straightforward x86-64 binary
  ELFCLASS32/EM_X86_64 - x32 (32 bit word size using x86-64 registers)

But we don't want to open any that don't match the category we
actually are.