print-libstubs: move to the new symbol loader/dumper implementation
This uses the same mechanism as libcapsule itself for loading symbols
from non-native DSO trees (under /host by default), so can be used
in the same environment as libcapsule itself is intended to operate
in.

Before this change the stub library tooling needed a different
(and incompatible) setup to libcapsule to work.