Make libcapsule check whether the proxy library is opening itself
We want the proxy library to open the ‘real’ target, not itsef.

Normally this is fine as the proxy is found by the linker and
the target is found by libcapsule under the prefix but if the
user is using LD_LIBRARY_PATH or LD_PRELOAD and/or no /prefix
we might need to take extra steps to prevent an infinite proxy
loop.